+++ example/child.html
@@ -4,8 +4,7 @@
 
     <title>Iframe Storage | Child</title>
 
-    <script type='text/javascript' src='../library/ift.js'></script>
-    <script type='text/javascript' src='../library/services/storage.js'></script>
+    <script type='text/javascript' src='../build/storage.js'></script>
     <script type='text/javascript'>
     ift.connect({
       trustedOrigins: ['http://127.0.0.1:8000']
