+++ example/parent.html
@@ -5,8 +5,7 @@
     <title>Iframe Storage | Parent</title>
 
     <script type='text/javascript' src='../bower_components/domready/ready.js'></script>
-    <script type='text/javascript' src='../library/ift.js'></script>
-    <script type='text/javascript' src='../library/services/storage.js'></script>
+    <script type='text/javascript' src='../build/storage.js'></script>
     <script type='text/javascript'>
     domready(function() {
       ift.connect({
