+++ library/util/isArray.js
@@ -1,0 +1,5 @@
+var toString = Object.prototype.toString;
+
+module.exports = Array.isArray || function(obj) {
+  return toString.call(obj) === '[object Array]';
+}
